+++ src/pipelines/js.rs
@@ -1,4 +1,4 @@
-//! CSS asset pipeline.
+//! JS asset pipeline.
 
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -7,11 +7,11 @@ use anyhow::{Context, Result};
 use nipper::Document;
 use tokio::task::JoinHandle;
 
-use super::{AssetFile, LinkAttrs, TrunkAssetPipelineOutput, ATTR_HREF};
+use super::{AssetFile, TrunkAssetPipelineOutput};
 use crate::config::RtcBuild;
 
-/// A CSS asset pipeline.
-pub struct Css {
+/// A JS asset pipeline.
+pub struct Js {
     /// The ID of this pipeline's source HTML element.
     id: usize,
     /// Runtime build config.
@@ -20,21 +20,18 @@ pub struct Css {
     asset: AssetFile,
 }
 
-impl Css {
-    pub const TYPE_CSS: &'static str = "css";
-
+impl Js {
     pub async fn new(
         cfg: Arc<RtcBuild>,
         html_dir: Arc<PathBuf>,
-        attrs: LinkAttrs,
+        src: Option<String>,
         id: usize,
     ) -> Result<Self> {
         // Build the path to the target asset.
-        let href_attr = attrs.get(ATTR_HREF).context(
-            r#"required attr `href` missing for <link data-trunk rel="css" .../> element"#,
-        )?;
+        let src_attr =
+            src.context(r#"required attr `src` missing for <script data-trunk .../> element"#)?;
         let mut path = PathBuf::new();
-        path.extend(href_attr.split('/'));
+        path.extend(src_attr.split('/'));
         let asset = AssetFile::new(&html_dir, path).await?;
         Ok(Self { id, cfg, asset })
     }
@@ -49,13 +46,13 @@ impl Css {
     #[tracing::instrument(level = "trace", skip(self))]
     async fn run(self) -> Result<TrunkAssetPipelineOutput> {
         let rel_path = crate::common::strip_prefix(&self.asset.path);
-        tracing::info!(path = ?rel_path, "copying & hashing css");
+        tracing::info!(path = ?rel_path, "copying & hashing js");
         let file = self
             .asset
             .copy(&self.cfg.staging_dist, self.cfg.filehash)
             .await?;
-        tracing::info!(path = ?rel_path, "finished copying & hashing css");
-        Ok(TrunkAssetPipelineOutput::Css(CssOutput {
+        tracing::info!(path = ?rel_path, "finished copying & hashing js");
+        Ok(TrunkAssetPipelineOutput::Js(JsOutput {
             cfg: self.cfg.clone(),
             id: self.id,
             file,
@@ -63,21 +60,21 @@ impl Css {
     }
 }
 
-/// The output of a CSS build pipeline.
-pub struct CssOutput {
+/// The output of a JS build pipeline.
+pub struct JsOutput {
     /// The runtime build config.
     pub cfg: Arc<RtcBuild>,
     /// The ID of this pipeline.
     pub id: usize,
-    /// Name the finalized output file.
+    /// Name of the finalized output file.
     pub file: String,
 }
 
-impl CssOutput {
+impl JsOutput {
     pub async fn finalize(self, dom: &mut Document) -> Result<()> {
-        dom.select(&super::trunk_id_selector(self.id))
+        dom.select(&super::trunk_script_id_selector(self.id))
             .replace_with_html(format!(
-                r#"<link rel="stylesheet" href="{base}{file}"/>"#,
+                r#"<script src="{base}{file}"/>"#,
                 base = &self.cfg.public_url,
                 file = self.file
             ));